I think I've got double diffusion working - at the very least, I can recreate Figure 4a from Large et al (diffusivity in the salt fingering regime). I'm going to turn this in to a case like the Bryan-Lewis mixing, where there are two columns being output so I can also test the diffusivity in the diffusive convection regime. Protex documentation is still a work in progress, so I'm not committing the change in the pdf just yet. Note that this change required modifications to the output routine (to output Rrho, the stratification parameter... needed because it is the x-axis in the plots).
Also, there was a bug in the Makefile where it was possible for the libcvmix.a library to get updated but then not linked in to the executable (make would think that the cvmix executable was up to date). I've fixed that by adding $(LIB_DIR)/libcvmix.a to the dependencies of $(EXE), and also added necessary $(INC_DIR)/___.mod files to the driver modules.



git-svn-id: https://cvmix.googlecode.com/svn/trunk@80 ec264166-b127-864d-6cb2-e100125a5c76