Modularized regression tests
I introduced reg_tests/common/ and put all the bash script that is used by
every regression test there. I think this is the first step towards having a
single testing script that takes the test you want to run as an argument, which
should make running the whole suite of tests significantly easier. For now,
this is important because it propogates the "--no_build" option from the
Bryan-Lewis test to tall the rest.

Automating the tidal test might be tough because the SVN link to the inputdata
repository requires validating the server certificate the first time you try to
download the data... maybe I can include a -inputdata option to point to an
existing checkout of the data instead of pulling it down every time.