Flexibility in where tidal test looks for data
Can now specify an inputdata directory with the -inputdata flag for the Simmons
tidal regression test (default is still $CVMIX/inputdata), this should let me
add tidal test to the automated suite. Required a little tweaking in how
namelists are handled because I need the tidal namelist file to point to the
correct directory for the data files... this is currently done with sed, which
is ugly but gets the job done.