Avoiding "empty object file" warning bit me
I tried to be smart about which compilers included cvmix_kinds_and_types.o in
libcvmix.a to avoid warnings when some compilers (basically all but PGI) create
an empty object file; this extra layer of complexity caused problems when
building CVMix as part of CESM on yellowstone with the PGI compiler. I don't
know why I was so interested in suppressing this particular warning message in
the first place, but I'll need to make a new CVMix tag to fix the POP build
issues.