We don't need buoyancy_cntr in cvmix_data_type
As Todd pointed out, we don't need the buoyancy_cntr field in cvmix_data_type,
it was only used by the stand-alone KPP driver and I/O module. I added an
optional "buoyancy_cntr" argument to the I/O routine and have the KPP driver
pass the buoyancy that needs to be written instead of pointing to it from
CVmix_vars.

I'm not planning on making a new tag just for this change, but it will
eventually find itself in v0.54-beta or the 1.0 release.