Add PP test for shear mixing, clean up other tests
Main update: I added the ability to produce data needed for PP81 Figure 4,
though I still need to write the NCL script to generate the plot. At this time,
there are no namelist variables to use with this test, it sets the parameters
needed for the 6 different runs and that's that.

Smaller fixes: Improved the cvmix_io.F90 code that is called when writing data
from an array of cvmix_data_type variables (the "multicol" write routine); this
affected the Bryan-Lewis driver (the only other reg_test that calls the
multiple column write). Then, while cleaning up the Bryan-Lewis code I realized
that the NCL file used to produce the plots were looking for output files that
no longer get created... so I cleaned that up as well. Plus there was a
misleading comment in the double diffusion driver.