Add support for Kelley mixing in diffusive regime
In the diffusive convection regime of double diffusion mixing, LMD94 compares
two parameterizations: Marmorino and Caldwell, 1976, and Kelley, 1988. The
recommendation is to only use MC76, which is all that was added to CVMix
initially... but I also added support for K88 to make it easier to reproduce
Figure 4b from LMD94. I added another double diffusion parameter named
"diff_conv_type", and the valid options are "MC76" (which is the default) and
"K88".

With this in place, I updated the ddiff reg test as well. The old test was not
set up in a very intuitive sense, but the new version should be a little easier
to follow. It sets up two columns, each of length 2*nlev; the first nlev levels
are in the salt fingering regime (1<=Rrho<=2) and the next nlev levels are in
the Diffusive Convection regime (0<Rrho<=1). The first column uses MC76 in the
latter regime, while the second column uses K88. Both columns use the same
parameterization in the salt fingering regime.

The change in column structure (one column that has both salt fingering and
diffusive convection instead of one column for each regime) meant that I could
change the stand-alone I/O routine to only output Rrho a single time rather
than needing to output it once per column.

I also upgraded to NCL version 6.2.1, and "isfilepresent()" has been changed;
this means I needed to update all of the plotting scripts to use "fileexists()"
instead, which means the NCL scripts will fail in older NCL versions.