New check_for_file() NCL routine
NCL 6.2.1 changed the output from the isfilepresent() routine - instead of
returning true if the file exists, the file must also be of a format that is
readable by addfile(). 6.2.1 also introduced the function fileexists() with the
same behavior as the old isfilepresent(), so I created version_consistency.ncl
which only contains check_for_file() -- if fileexists() exists, it is used to
check for the file, otherwise isfilepresent() is used.

It's a little clunky, but it's better than requiring NCL 6.2.1 or later to
generate plots.