Add minVtSqr parameter (default = 1e-10 m^2/s^2)
To prevent division by zero if the unresolved shear is 0, we impose a minimum
value on Vtsqr. This parameter can be set in cvmix_init_kpp but the default
is 1e-10 m^2/s^2.

Note that there is still a check in compute_bulk_Richardson to see if

|V-Vr|^2 + Vt^2 = 0

If so, Ri_bulk = (1-0.5*surf_layer_ext)*d*(B-Br)*1e10

This 1e10 is hardcoded, not a parameter.