Improve stand-alone build tie-in to netCDF
I needed to use homebrew to install netCDF on my laptop for a different project
and the way homebrew distributed the library and include files didn't play nice
with the old system... so now the first thing CVMix does is look for nc-config
to determine both the include directory and the linking flags. If that
information isn't available, it reverts back to the old build.