More protex updates
Found a better protex script to use as a base (courtesy of GEOS-Chem at Harvard
- I believe this is the group that provided the original f90pdf script that we
were previously using). Made a few tweaks to get it working with modern perl
(based on changes from the previous protex script I had found), and also
cleaned up a lot of the protex comments in the source code itself.